fix: Remove overloads and move helper methods to utils
